requirements.txt
py-cord==2.5.0
tortoise-orm==0.21.3
asyncpg==0.29.0